# Implement FasterAccurateLocalTimeSource