This TimeSource is close to the performance of TimeSource.AccurateUtc but returns local time

Here is benchmark results using [BenchmarkDotNet](https://github.com/dotnet/BenchmarkDotNet)
[Source code of benchmark](https://gist.github.com/Encamy/51a789bf00450aa117596ecd1e1a1fa3)

``` ini

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17763.475 (1809/October2018Update/Redstone5)
Intel Core i5-3570 CPU 3.40GHz (Ivy Bridge), 1 CPU, 4 logical and 4 physical cores
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3324.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3324.0


```
|               Method |       Mean |     Error |    StdDev |
|--------------------- |-----------:|----------:|----------:|
|            FastLocal |   5.784 ns | 0.0340 ns | 0.0318 ns |
|              FastUtc |   5.246 ns | 0.0303 ns | 0.0284 ns |
|        AccurateLocal | 215.149 ns | 0.8370 ns | 0.7830 ns |
|          AccurateUtc | 147.001 ns | 0.7035 ns | 0.6581 ns |
| AccurateAndFastLocal | 152.786 ns | 0.7137 ns | 0.6676 ns |