@304NotModified 2019-05-08
Cool! Thanks!

@snakefoot as you're better with performance wise code, any opinion on this?