@snakefoot 2019-05-08
Benchmarks are usually the right way to measure performance. If they improve then it is usually good.

But when dealing with local-time and utc-time. Then one need to consider Daylight-Saving-Time (DST) this means that local-time can jump in time-difference compared to utc-time.

The hack with saving the time-difference in ticks between local-time and utc-time needs to be updated atleast every second.

Because the Time-source has to be highly concurrent then multiple threads will attempt to "refresh" the time-difference-cache. The refresh caching-logic needs to cater for this, and update two properties in one operations:

- Current Time Difference Ticks
- Last Updated Time Difference

The most optimal threadsafe way is probably using Interlocked-methods. These usually prefer primitive values (or references). So "Last Updated Time Difference" should probably be stored as DateTime.UtcNow.Ticks.