@Encamy 2019-05-10
I understand that DST is very important, but what about countries where it [does not exist](https://en.wikipedia.org/wiki/Daylight_saving_time_by_country)? If cached time will be updated every second, the results will be as follows:

|              Method |       Mean |     Error |    StdDev |
|-------------------- |-----------:|----------:|----------:|
|           FastLocal |   5.965 ns | 0.0975 ns | 0.0912 ns |
|             FastUtc |   5.155 ns | 0.0494 ns | 0.0438 ns |
|       AccurateLocal | 210.057 ns | 1.4504 ns | 1.2857 ns |
|         AccurateUtc | 144.635 ns | 2.3186 ns | 2.1689 ns |
| FasterAccurateLocal_updateEachSecond | 197.228 ns | 2.1071 ns | 1.9710 ns |
| FasterAccurateLocal_updateEachMinute | 183.867 ns | 0.8148 ns | 0.7621 ns |

Seems like there is no point in new TimeSource if it will update cache data.

I did not create a commit because perfomance with updating cache data is not great, here is a [code](https://gist.github.com/Encamy/d736fffd11fba2e7bc8f03c6c3b9cf27)

What about creating TimeSource that will not check for DST? For example boolean flag in constructor (this will be fast for users without DST and a bit faster than AccurateLocal for users with DST) or rename this time source to AccurateLocal_NoDST?