@304NotModified 2019-05-10
> What about creating TimeSource that will not check for DST? For example boolean flag in constructor (this will be fast for users without DST and a bit faster than AccurateLocal for users with DST) or rename this time source to AccurateLocal_NoDST?

I'm not sure if someone is looking for Local time with NoDST, then use UTC IMO