@snakefoot 2019-05-11
> I did not create a commit because perfomance with updating cache data is not great, here is a code

You are doing many operations for checking if 1 second has passed. Why not this:

```c#
private DateTime GetTime()
{
  DateTime utc = DateTime.UtcNow;
  int ticksOffset = _ticksOffset;
  long lastUpdatedTicks = Interlocked.Read(ref _lastUpdated);
  long deltaTicks = utc.Ticks - lastUpdatedTicks;
  if (deltaTicks > TimeSpan.TicksPerSecond || deltaTicks < -TimeSpan.TicksPerSecond)
  {
      _ticksOffset = ticksOffset = (DateTime.Now - utc).Ticks;
      _lastUpdated = utc.Ticks;
  }
  long ticks = utc.Ticks + ticksOffset;
  return new DateTime(ticks);
}
```