@Encamy 2019-05-12
Thanks, @snakefoot!
Your implementation is better, but why do we need local `long ticksOffset`?
Why we can't use updated cache offset? Is it because of a possible data race in multithreaded environment?

Here is new benchmark. It looks amazing

|              Method |       Mean |     Error |    StdDev |
|-------------------- |-----------:|----------:|----------:|
|           FastLocal |   5.205 ns | 0.0213 ns | 0.0189 ns |
|             FastUtc |   5.202 ns | 0.0266 ns | 0.0249 ns |
|       AccurateLocal | 213.456 ns | 0.5400 ns | 0.5051 ns |
|         AccurateUtc | 144.859 ns | 0.4159 ns | 0.3891 ns |
| FasterAccurateLocal | 158.777 ns | 0.4778 ns | 0.4469 ns |