@304NotModified 2019-05-12
PS  On appveyor we build more platforms than locally (xamarin, silverlight).