@snakefoot 2019-05-12
> Your implementation is better, but why do we need local long ticksOffset?

I just like local copies when concurrency is in play. Maybe more can be be squezed out with not using Ticks-property constantly.

```c#
private DateTime GetTime()
{
  long utcTicks = DateTime.UtcNow.Ticks;
  long deltaTicks = utcTicks - Interlocked.Read(ref _lastUpdated);
  long ticksOffset = _ticksOffset;
  if (deltaTicks > TimeSpan.TicksPerSecond || deltaTicks < -TimeSpan.TicksPerSecond)
  {
      _ticksOffset = ticksOffset = (DateTime.Now.Ticks - utcTicks);
      _lastUpdated = utcTicks;
  }
  return new DateTime(utcTicks + ticksOffset, DateTimeKind.Local);
}
```

Also ensures the DateTimeKind is correct for the Local-time (Instead of unspecified).