@304NotModified 2019-05-24
Thanks for fixes!

Would it make sense to replace the AccurateLocal with FasterAccurateLocal?