@snakefoot 2019-05-24
I would probably name it `FastAccurateLocal` it matches `FastLocal` (Instead of `FasterAccurateLocal`)