@Encamy 2019-05-28
I think that it can't be named `AccurateLocal`, bacause we use some caching here. If users want to use really accurate and reliable time source, they should use `AccurateLocal` in the way in which it is implemented now.
I think we can name it `FastAccurateLocal` since it is both fast and accurate. Do you agree?
Should I add unit tests?