@304NotModified 2019-05-28
Sorry to say, but if it isn't accurate, i'm doubting if we should add this to NLog. 

I'm not sure if the burden of another timesource is the improvement worth (~25% faster)

@snakefoot what do you think?