@Encamy 2019-05-28
This time source is accurate. While I was using it I didn't notice any problems
But I think that not everyone will be happy when they notice that their time source has been changed.