@snakefoot 2019-05-28
> @snakefoot what do you think?

I prefer not to replace the existing `AccurateLocal`. But `FastAccurateLocal` will never be faster than `AccurateUtc`. Have no opinion about adding another timesource-type. Normally use AccurateUtc myself.