@304NotModified 2019-05-28
> This time source is accurate.

OK clear :)

FastAccurateLocal  is then fine to me. UTC is always faster than non-UTC 

@Encamy could you please update https://github.com/NLog/NLog/wiki/Time-Source? Thanks