@304NotModified 2019-05-28
OK thanks, so next steps are tests :)