@snakefoot 2019-05-28
@Encamy Would it be possible for you to apply the changes from my comment: https://github.com/NLog/NLog/pull/3378#issuecomment-491620318 (Avoid the Ticks-lookup)