@Encamy 2019-05-28
I've applied changes from https://github.com/NLog/NLog/pull/3378#issuecomment-491620318 (sorry, I didn't notice it earlier)

I noticed that in [wiki page](https://github.com/NLog/NLog/wiki/Time-Source) results of AccurateUtc are different from my tests. Should I update this value or not?
Here is last test run:

|            Method |       Mean |     Error |    StdDev |     Median |
|------------------ |-----------:|----------:|----------:|-----------:|
|         FastLocal |   5.854 ns | 0.0216 ns | 0.0191 ns |   5.847 ns |
|           FastUtc |   5.839 ns | 0.0197 ns | 0.0165 ns |   5.834 ns |
|     AccurateLocal | 190.481 ns | 0.5205 ns | 0.4869 ns | 190.631 ns |
|       AccurateUtc | 137.258 ns | 0.4862 ns | 0.4310 ns | 137.423 ns |
| FastAccurateLocal | 150.147 ns | 0.1401 ns | 0.1311 ns | 150.150 ns |

I will make unit tests as soon as possible