@snakefoot 2019-05-28
> Here is last test run:

Is that on 32 bit / 64 bit ? Net472 / NetCore22 ?